solutions/advanced_workshop/04-multi-agent-workflow-dapr-workflows/services/wizard/app.py
from dapr_agents import AssistantAgent
from dapr_agents.llm import HFHubChatClient
from dotenv import load_dotenv
import asyncio
import logging

async def main():
    try:
        llm_client = HFHubChatClient(
            model="microsoft/Phi-3-mini-4k-instruct"
        )
        wizard_service = AssistantAgent(
            role="Wizard",
            name="Gandalf",
            goal="Guide the Fellowship with wisdom and strategy, using magic and insight to ensure the downfall of Sauron.",
            instructions=[
                "Speak like Gandalf, with wisdom, patience, and a touch of mystery.",
                "Provide strategic counsel, always considering the long-term consequences of actions.",
                "Use magic sparingly, applying it when necessary to guide or protect.",
                "Encourage allies to find strength within themselves rather than relying solely on your power.",
                "Respond concisely, accurately, and relevantly, ensuring clarity and strict alignment with the task.",
            ],
            message_bus_name="messagepubsub",
            state_store_name="workflowstatestore",
            state_key="workflow_state",
            agents_registry_store_name="agentstatestore",
            agents_registry_key="agents_registry",
            llm=llm_client
        )

        await wizard_service.start()
    except Exception as e:
        print(f"Error starting service: {e}")


if __name__ == "__main__":
    load_dotenv()

    logging.basicConfig(level=logging.INFO)

    asyncio.run(main())
